+++ Source/Archetypes/chemist.js
@@ -4,4 +4,4 @@ module.exports = new Archetype("Chemist")
 	.setElement("Water")
 	.setPredictType("Health")
 	.setDescription("Able to and assess combatant modifiers and hp levels, the Chemist excels at managing party and enemy health.")
-	.setSignatureEquipment(["Sickle", "Unfinished Potion"]);
+	.setSignatureEquipment(["Sickle", "Potion Kit"]);