+++ Source/Commands/delve.js
@@ -16,8 +16,8 @@ const options = [
 ];
 module.exports = new Command(id, "Start a new adventure", false, false, options);
 
-let DESCRIPTORS = ["Shining", "New", "Dusty", "Old", "Floating", "Undersea", "Future", "Intense"];
-let LOCATIONS = ["Adventure", "Castle", "Labyrinth", "Ruins", "Plateau", "Dungeon", "Maze", "Fortress", "Dream"];
+const DESCRIPTORS = ["Shining", "New", "Dusty", "Old", "Floating", "Undersea", "Future", "Intense"];
+const LOCATIONS = ["Pyramid", "Adventure", "Castle", "Labyrinth", "Ruins", "Plateau", "Dungeon", "Maze", "Fortress", "Dream"];
 
 module.exports.execute = (interaction) => {
 	// Start a new adventure
@@ -30,6 +30,7 @@ module.exports.execute = (interaction) => {
 				prerollBoss("Final Battle", adventure);
 				prerollBoss("Artifact Guardian", adventure);
 
+				//TODO #469 generate adventure name using labyrinth name
 				let elementPool = elementsList();
 				let pickedElement = elementPool[generateRandomNumber(adventure, elementPool.length, "general")];
 				adventure.setName(`${DESCRIPTORS[generateRandomNumber(adventure, DESCRIPTORS.length, "general")]} ${LOCATIONS[generateRandomNumber(adventure, LOCATIONS.length, "general")]} of ${pickedElement}`)