+++ Source/equipment/_equipmentDictionary.js
@@ -1,36 +1,6 @@
-const Adventure = require("../../Classes/Adventure.js");
-const { generateRandomNumber } = require("../../helpers.js");
 const { getEmoji } = require("../elementHelpers.js");
 
-const allEquipment = {};
-const equipmentDrops = {
-	Earth: {
-		"0": [],
-		"1": [],
-		"2": []
-	},
-	Wind: {
-		"0": [],
-		"1": [],
-		"2": []
-	},
-	Water: {
-		"0": [],
-		"1": [],
-		"2": []
-	},
-	Fire: {
-		"0": [],
-		"1": [],
-		"2": []
-	},
-	Untyped: {
-		"-1": [],
-		"0": [],
-		"1": [],
-		"2": []
-	}
-};
+const EQUIPMENT = {};
 
 for (const file of [
 	"-punch.js",
@@ -84,10 +54,7 @@ for (const file of [
 	"midasstaff-base.js",
 	"midasstaff-soothing.js",
 	"midasstaff-accelerating.js",
-	"potion-base.js",
-	"potion-earthen.js",
-	"potion-watery.js",
-	"potion-windy.js",
+	"potionkit-base.js",
 	"scythe-base.js",
 	"scythe-lethal.js",
 	"scythe-piercing.js",
@@ -116,16 +83,15 @@ for (const file of [
 	"warhammer-piercing.js"
 ]) {
 	const equip = require(`./${file}`);
-	allEquipment[equip.name] = equip;
-	equipmentDrops[equip.element][equip.tier.toString()].push(equip.name);
-}
+	EQUIPMENT[equip.name] = equip;
+};
 
 /** Checks if a type of equipment with the given name exists
  * @param {string} equipmentName
  * @returns {boolean}
  */
 exports.equipmentExists = function (equipmentName) {
-	return equipmentName in allEquipment;
+	return equipmentName in EQUIPMENT;
 }
 
 /** Fetch a single default property from an existing type of equipment
@@ -135,7 +101,7 @@ exports.equipmentExists = function (equipmentName) {
  */
 exports.getEquipmentProperty = function (equipmentName, propertyName) {
 	if (exports.equipmentExists(equipmentName)) {
-		return allEquipment[equipmentName][propertyName];
+		return EQUIPMENT[equipmentName][propertyName];
 	} else {
 		console.error("Fetching property from illegal equipment: " + equipmentName);
 	}
@@ -163,14 +129,4 @@ exports.buildEquipmentDescription = function (equipmentName, buildFullDescriptio
 			return description.split("\n")[0].replace(/\*/g, "");
 		}
 	}
-}
-
-/** Filters by party element pool and given tier, then rolls a random equipment's name
- * @param {"0" | "1" | "2"} tier
- * @param {Adventure} adventure
- * @returns {string} the name of the dropped equipment
- */
-exports.rollEquipmentDrop = function (tier, adventure) {
-	const pool = adventure.getElementPool().reduce((pool, element) => pool.concat(equipmentDrops[element][tier]), []);
-	return pool[generateRandomNumber(adventure, pool.length, "general")];
 }