+++ ChangeLog.md
@@ -15,7 +15,9 @@
    12. Strength Spinach
    13. Watery Potion
    14. Windy Potion
+   15. Block Potion
 - Reworked *Urgent* weapon variant: it now adds priority to the move
+- Reworked *Unfinished Potion* to *Potion Kit*: it now adds a random "potion" consumable to loot
 - Renamed *Swift* weapon variant to *Accelerating*
 
 #### Prophets of the Labyrinth Version 0.9.0: